Move RunMojo follow tests to a dedicated RunMojoTest class
Per review feedback: RunMojo assertions do not belong in StartMojoTest.
Extract runFollowDefaultsToTrue / runHonoursConfiguredFollow into a new
RunMojoTest and keep StartMojoTest focused on docker:start.
