+++ src/test/java/io/fabric8/maven/docker/StartMojoTest.java
@@ -6,8 +6,8 @@ import org.junit.jupiter.api.Test;
 
 /**
  * Verifies that the {@code follow} log-following flag can be driven from the {@code <follow>}
- * configuration element (and not only the {@code docker.follow} system property), while keeping
- * the historical defaults: {@code docker:start} does not follow, {@code docker:run} does. See #1797.
+ * configuration element (and not only the {@code docker.follow} system property), and that
+ * {@code docker:start} does not follow logs unless configured. See #1797.
  */
 class StartMojoTest {
 
@@ -27,23 +27,5 @@ class StartMojoTest {
 
         mojo.follow = Boolean.FALSE;
         Assertions.assertFalse(mojo.followLogs());
-    }
-
-    @Test
-    @DisplayName("docker:run follows logs by default (default true)")
-    void runFollowDefaultsToTrue() {
-        Assertions.assertTrue(new RunMojo().followLogs());
-    }
-
-    @Test
-    @DisplayName("docker:run honours the <follow> configuration element")
-    void runHonoursConfiguredFollow() {
-        RunMojo mojo = new RunMojo();
-
-        mojo.follow = Boolean.FALSE;
-        Assertions.assertFalse(mojo.followLogs());
-
-        mojo.follow = Boolean.TRUE;
-        Assertions.assertTrue(mojo.followLogs());
     }
 }
